chore: move folders out of ui folder
