+++ log-viewer/src/components/BadgeBase.ts
