+++ log-viewer/src/components/CallStack.ts
@@ -4,9 +4,9 @@
 import { LitElement, css, html } from 'lit';
 import { customElement, property } from 'lit/decorators.js';
 
-import type { LogEvent } from '../../core/log-parser/LogEvents.js';
-import { goToRow } from '../../features/call-tree/components/CalltreeView.js';
-import { DatabaseAccess } from '../../features/database/services/Database.js';
+import type { LogEvent } from '../core/log-parser/LogEvents.js';
+import { goToRow } from '../features/call-tree/components/CalltreeView.js';
+import { DatabaseAccess } from '../features/database/services/Database.js';
 import { globalStyles } from '../styles/global.styles.js';
 
 @customElement('call-stack')
