+++ log-viewer/src/components/VsIconCheckbox.ts
