+++ log-viewer/src/components/datagrid-filter-bar.ts
