+++ log-viewer/src/features/database/components/DatabaseSOQLDetailPanel.ts
@@ -4,8 +4,8 @@
 import { LitElement, css, html } from 'lit';
 import { customElement, property } from 'lit/decorators.js';
 
-import '../../../ui/components/CallStack.js';
-import { globalStyles } from '../../../ui/styles/global.styles.js';
+import '../../../components/CallStack.js';
+import { globalStyles } from '../../../styles/global.styles.js';
 import '../../soql/components/SOQLLinterIssues.js';
 
 @customElement('db-soql-detail-panel')
