+++ log-viewer/src/features/database/components/DatabaseView.scss
@@ -1,4 +1,4 @@
-@import '.././../../ui/datagrid/style/DataGrid.scss';
+@import '../../../tabulator/style/DataGrid.scss';
 
 .row__details-container {
   height: 100%;
