+++ log-viewer/src/features/database/components/DatabaseView.ts
@@ -5,9 +5,9 @@
 import { LitElement, css, html } from 'lit';
 import { customElement, property, state } from 'lit/decorators.js';
 
+import '../../../components/CallStack.js';
 import type { ApexLog } from '../../../core/log-parser/LogEvents.js';
-import '../../../ui/components/CallStack.js';
-import { globalStyles } from '../../../ui/styles/global.styles.js';
+import { globalStyles } from '../../../styles/global.styles.js';
 import './DMLView.js';
 import './DatabaseSOQLDetailPanel.js';
 import './DatabaseSection.js';
