+++ log-viewer/src/features/soql/components/SOQLLinterIssues.ts
@@ -5,7 +5,7 @@ import { LitElement, css, html, type PropertyValues, type TemplateResult } from 
 import { customElement, property, state } from 'lit/decorators.js';
 
 import type { SOQLExecuteBeginLine } from '../../../core/log-parser/LogEvents.js';
-import { globalStyles } from '../../../ui/styles/global.styles.js';
+import { globalStyles } from '../../../styles/global.styles.js';
 import { DatabaseAccess } from '../../database/services/Database.js';
 import {
   SEVERITY_TYPES,
