+++ log-viewer/src/styles/codicon.css
