+++ log-viewer/src/styles/global.styles.ts
