+++ log-viewer/src/styles/notification.styles.ts
