+++ log-viewer/src/styles/skeleton.styles.ts
