+++ log-viewer/src/tabulator/dataaccessor/Number.ts
