+++ log-viewer/src/tabulator/editors/MinMax.css
