+++ log-viewer/src/tabulator/editors/MinMax.ts
