+++ log-viewer/src/tabulator/filters/MinMax.ts
