+++ log-viewer/src/tabulator/format/Number.ts
