+++ log-viewer/src/tabulator/format/Progress.css
