+++ log-viewer/src/tabulator/format/Progress.ts
