+++ log-viewer/src/tabulator/format/ProgressComponent.ts
