+++ log-viewer/src/tabulator/format/ProgressMS.ts
