+++ log-viewer/src/tabulator/groups/GroupCalcs.ts
