+++ log-viewer/src/tabulator/style/DataGrid.scss
